requirements.txt
streamlit
langchain
langchain-core
langchain-openai
langchain-groq
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-experimental
sentence-transformers
python-dotenv
groq
notebook
pandas
faiss-cpu
tiktoken
tabulate
